requirements.txt
slackclient==2.5.0
PyGithub==1.46